pyproject.toml
[project]
name = "discord-utils-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Stefano Di Pasquale",email = "stefaniscion@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py (>=2.6.3,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
